requirements.txt
Django
django-crispy-forms==1.14.0
mssql-django
python-decouple
whitenoise
O365
